refactor: folders and theme provider
