+++ package.json
@@ -16,13 +16,13 @@
     "@mdx-js/react": "3.1.0",
     "clsx": "2.1.1",
     "modern-normalize": "3.0.1",
-    "next-themes": "0.4.6",
     "posthog-js": "1.215.6",
     "react": "19.0.0",
     "react-aria-components": "1.8.0",
     "react-dom": "19.0.0",
     "remark-gfm": "4.0.1",
-    "tuono": "0.19.6"
+    "tuono": "0.19.7",
+    "usehooks-ts": "3.1.1"
   },
   "devDependencies": {
     "@eslint/js": "9.21.0",
