+++ src/components/MdxProvider/MdxProvider.tsx
@@ -1,4 +1,4 @@
-import type { JSX, ReactNode } from 'react'
+import { useMemo, type JSX, type ReactNode } from 'react'
 import { MDXProvider } from '@mdx-js/react'
 
 interface MdxProviderProps {
@@ -6,24 +6,29 @@ interface MdxProviderProps {
 }
 
 export function MdxProvider({ children }: MdxProviderProps): JSX.Element {
-  return (
-    <MDXProvider
-      components={{
-        a: (props) => <a {...props} />,
-        h1: (props) => <h1 {...props} />,
-        h2: (props) => <h2 {...props} />,
-        h3: (props) => <h3 {...props} />,
-        h4: (props) => <h4 {...props} />,
-        h5: (props) => <h5 {...props} />,
-        h6: (props) => <h6 {...props} />,
-        strong: (props) => <strong {...props} />,
-        p: (props) => <p {...props} />,
-        hr: () => <hr />,
-        pre: (props) => <pre {...props} style={{ width: '100px' }} />,
-        code: (props) => <pre {...props} />,
-      }}
-    >
-      {children}
-    </MDXProvider>
+  const Provider = useMemo(
+    () => (
+      <MDXProvider
+        components={{
+          a: (props) => <a {...props} />,
+          h1: (props) => <h1 {...props} />,
+          h2: (props) => <h2 {...props} />,
+          h3: (props) => <h3 {...props} />,
+          h4: (props) => <h4 {...props} />,
+          h5: (props) => <h5 {...props} />,
+          h6: (props) => <h6 {...props} />,
+          strong: (props) => <strong {...props} />,
+          p: (props) => <p {...props} />,
+          hr: () => <hr />,
+          pre: (props) => <pre {...props} style={{ width: '100px' }} />,
+          code: (props) => <pre {...props} />,
+        }}
+      >
+        {children}
+      </MDXProvider>
+    ),
+    [],
   )
+
+  return Provider
 }
